refactor: apis 폴더 구조 변경
[#74]
